refactor: split math-traversal-normalization.ts into focused modules
Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/7943cca9-ab9b-4045-ad5d-bd79943a9f66
